+++ src/filter/env_filter/README.md
@@ -1,3 +1,3 @@
 # RUST_LOG Environment Variable Filter
 
-This filter is a fork of [env_filter](https://crates.io/crates/env_filter), with adaption to use this crate's logging types.
+This filter is a fork of [env_filter](https://crates.io/crates/env_filter), with adaptation to use this crate's logging types.
